refactor code architecture
